refactor hooks into separate files
