+++ src/window_main/components/collection/filters.tsx
@@ -6,8 +6,8 @@ import { CARD_RARITIES, MANA } from "../../../shared/constants";
 import db from "../../../shared/database";
 import { RaritySymbol, SetSymbol, TypeSymbol } from "../misc/display";
 import { BinaryColumnFilter, BinaryFilterValue } from "../tables/filters";
-import { useMultiSelectFilter } from "../tables/hooks";
 import { MultiSelectFilterProps } from "../tables/types";
+import { useMultiSelectFilter } from "../tables/useMultiSelectFilter";
 import { CardsData } from "./types";
 
 export function InBoostersColumnFilter(props: {
